pyproject.toml
[project]
name = "splitsilly"
version = "0.1.0"
description = ""
authors = [{ name = "Patrick Gingras", email = "775.pg.12@gmail.com" }]
requires-python = ">=3.13,<3.14"  # FIXME: update when ddtrace supports 3.14
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "django>=4.2.7,<5",
    "django-extensions>=3.2.3,<4",
    "ipython>=8.17.2,<9",
    "python-dotenv>=1.0.0,<2",
    "gunicorn>=21.2.0,<22",
    "sentry-sdk[django]>=1.40.0,<2",
    "ddtrace>=2.6.0,<3",
    "django-timezone-field>=6.1.0,<7",
    "yarl>=1.9.4,<2",
    "huey>=2.5.0,<3",
    "django-stubs-ext>=5.2.9",
]

[dependency-groups]
dev = [
    "pytest>=7.4.3,<8",
    "pytest-django>=4.7.0,<5",
    "pre-commit>=3.5.0,<4",
    "ruff>=0.14.14",
    "django-stubs[compatible-mypy]>=5.2.9",
    "mypy>=1.19.1",
]

[tool.uv]

[tool.uv.build-backend]
module-root = ""

[build-system]
requires = ["uv_build>=0.9.4,<0.10.0"]
build-backend = "uv_build"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "splitsilly.settings"

[tool.mypy]
python_version = "3.13"
strict = true
warn_unused_configs = true
follow_imports = "normal"
ignore_missing_imports = false
plugins = ["mypy_django_plugin.main"]
exclude = "tests"

[tool.django-stubs]
django_settings_module = "splitsilly.settings"
